Make the event source class public so the reflection data is correctly generated.
